backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
openpyxl==3.1.2
python-dateutil==2.8.2
